Muevo logica dentro del form a una clase aparte, que se usa en la view
